상품(숙박, 강습)을 조건으로 검색해서 나열해 주는 페이지 작성함.
다중 조건으로 검색 가능함.
기타 오류 등 수정됨.
평점 int → float
문의 내용 줄바꿈을 <br>처리시킴
roomManage.jsp 하단 footer 및 닫는 태그 미삽입 고침

+++ final_project/src/main/java/kr/co/lesson/model/vo/Lesson.java
@@ -23,7 +23,7 @@ public class Lesson {
 	private int lessonPrice;			// 강습(상품) 가격
 	private String lessonInfo;			// 게시글에서 내용(content)이라고 인지하면 됨.
 	private String lessonInfoPic;		// 첨부사진. 본문 첨부가 아니라 첨부파일임.
-	private String lessonTeacher;		// 강사이름. MEMBER_NAME을 가져오는게 아니라, 판매자가 임의로 적는 값		
+	private String lessonTeacher;		// 강사이름. MEMBER_NAME을 가져오는게 아니라, 판매자가 임의로 적는 값
 	private int lessonMaxNo;			// 수업 당 정원. 최대재고량으로 인지하면 됨.
 	private String lessonStartTime;		// "HH:MM " 시작시간만, default(선택시간) 로우  하나 더
 	private String lessonEndTime;		// "HH:MM " 종료시간만
@@ -32,7 +32,7 @@ public class Lesson {
 											0 = 판매중지
 										   -1 = 승인대기중. DB에는 insert됐지만, 고객들에게 노출되지는 않는 상품(게시물)
 	  									   -2 = 승인 반려됨. DB에는 insert됐지만, 고객들에게 노출되지는 않는 상품(게시물)	*/
-	private int lessonScore;			// 별평점후기 Table (RIVEIW)이 변경될때마다, oracle에서 trigger가 작동해서 LESSON 테이블의 LESSON_SCORE를 갱신
+	private float lessonScore;			// 별평점후기 Table (RIVEIW)이 변경될때마다, oracle에서 trigger가 작동해서 LESSON 테이블의 LESSON_SCORE를 갱신
 	private String writer;				// 상품등록자의 MEMBER_ID (글쓴이)
 	private String rejectComment;		// houseStatus -1인 게시물을 관리자가 반려해서 -2로 상태를 바꿀 때, 사유를 입력하는 란
